\name{Compositional-package}
\alias{Compositional-package}
\docType{package}
\title{
This is an R package that provides methods for statistical analysis of compositional data
}

\description{
Many types of regression and discriminant analysis, contour plots, parameter estimation and more functions are included. As a general rule, keep in mind that whenever
the log-ratio transformation (additive or isometric) is to be used the compositional data must have no zero values. For all other cases, feel free
to have zeros.
}

\details{
\tabular{ll}{
Package: \tab Compositional\cr
Type: \tab Package\cr
Version: \tab 1\cr
Date: \tab 2016-02-05\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris <mtsagris@yahoo.gr> and Giorgos Athineou <athineou@csd.uoc.gr>
}

\section{Maintainers }{
Michail Tsagris <mtsagris@yahoo.gr> and Giorgos Athineou <athineou@csd.uoc.gr>
}

\references{
Aitchison J. (1986). The statistical analysis of compositional data. Chapman & Hall.
}

\note{
Acknowledgments:

Professor Andy Wood and Dr Simon Preston from the university of Nottingham are highly appreciated for being my supervisors during my PhD in compositional data analysis.
}

\keyword{ Compositional data }
\keyword{ Regression }
\keyword{ Contour plots }
\keyword{ Discriminant analysis }
\keyword{ MLE }


