\name{Compositional-package}
\alias{Compositional-package}
\docType{package}
\title{
Compositional Data Analysis
}

\description{
A Collection of Functions for Compositional Data Analysis.
}

\details{
\tabular{ll}{
Package: \tab Compositional\cr
Type: \tab Package\cr
Version: \tab 6.9 \cr
Date: \tab 2024-07-08\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}, Giorgos Athineou <gioathineou@gmail.com>,
Abdulaziz Alenazi <a.alenazi@nbu.edu.sa> and Christos Adam \email{pada4m4@gmail.com}.
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>
}

\references{
Aitchison J. (1986). The statistical analysis of compositional data. Chapman & Hall.
}

\note{
Acknowledgments:

Michail Tsagris would like to express his acknowledgments to Professor Andy Wood and Professor Simon
Preston from the university of Nottingham for being his supervisors during his PhD in compositional data analysis.

We would also like to express our acknowledgments to Profesor Kurt Hornik (and also the rest of the R core team) for his help with this package.

Manos Papadakis, undergraduate student in the department of computer science, university of Crete, is also acknowledged for his programming tips.

Ermanno Affuso from the university of South Alabama suggested that I have a default value in the function \code{\link{mkde}}.

Van Thang Hoang from Hasselt university spotted a bug in the function \code{\link{js.compreg}}.

Claudia Wehrhahn Cortes spotted a bug in the function \code{\link{diri.reg}}.

Philipp Kynast from Bruker Daltonik GmbH found a mistake in the function \code{\link{mkde}} which is now fixed.

Jasmine Heyse from the university of Ghent spotted a bug in the function \code{\link{kl.compreg}} which is now fixed.

Magne Neby suggested to add names in the covariance matrix of the divergence based regression models.

John Barry from the Centre for Environment, Fisheries, and Aquaculture Science (UK) suggested that I should add more explanation in the function \code{\link{diri.est}}. I hope it is clearer now.

Charlotte Fabri and Laura Byrne spotted a possible problem in the function \code{\link{zadr}}.

Levi Bankston found a bug in the bootstrap version of the function \code{\link{kl.compreg}}.

Sucharitha Dodamgodage suggested to add an extra case in the function \code{\link{dirimean.test}}.

Loic Mangnier found a bug in the function \code{\link{lc.glm}} which is now fixed and also became faster.
}

\keyword{ Compositional data }
\keyword{ Regression }
\keyword{ Contour plots }
\keyword{ Discriminant analysis }
\keyword{ MLE }


