\name{Compositional-package}
\alias{Compositional-package}
\docType{package}
\title{
Compositional Data Analysis
}

\description{
A collection of functions for compositional data analysis.
}

\details{
\tabular{ll}{
Package: \tab Compositional\cr
Type: \tab Package\cr
Version: \tab 2.2 \cr
Date: \tab 2016-12-06\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris <mtsagris@yahoo.gr> and Giorgos Athineou <athineou@csd.uoc.gr>
}

\section{Maintainers }{
Michail Tsagris <mtsagris@yahoo.gr>
}

\references{
Aitchison J. (1986). The statistical analysis of compositional data. Chapman & Hall.
}

\note{
Acknowledgments:

Michail Tsagris would like to express his acknowledgments to Professor Andy Wood and Dr Simon Preston from the university of Nottingham for being his supervisors during his PhD in compositional data analysis. We would also like to express our acknowledgments to Profesor Kurt Hornik (and also the rest of the R core team) for his help with this package. Manos Papadakis, undergraduate student in the department of computer science, university of Crete, is also acknowledged for his programming tips.
}

\keyword{ Compositional data }
\keyword{ Regression }
\keyword{ Contour plots }
\keyword{ Discriminant analysis }
\keyword{ MLE }


