\name{Compositional-package}
\alias{Compositional-package}
\docType{package}
\title{
Compositional Data Analysis
}

\description{
A collection of functions for compositional data analysis.
}

\details{
\tabular{ll}{
Package: \tab Compositional\cr
Type: \tab Package\cr
Version: \tab 1.3 \cr
Date: \tab 2016-05-27\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris <mtsagris@yahoo.gr> and Giorgos Athineou <athineou@csd.uoc.gr>
}

\section{Maintainers }{
Michail Tsagris <mtsagris@yahoo.gr>
}

\references{
Aitchison J. (1986). The statistical analysis of compositional data. Chapman & Hall.
}

\note{
Acknowledgments:

Professor Andy Wood and Dr Simon Preston from the university of Nottingham are highly appreciated for being my supervisors during my PhD in compositional data analysis.
}

\keyword{ Compositional data }
\keyword{ Regression }
\keyword{ Contour plots }
\keyword{ Discriminant analysis }
\keyword{ MLE }


