\name{Compositional-package}
\alias{Compositional-package}
\docType{package}
\title{
Compositional Data Analysis
}

\description{
A collection of functions for compositional data analysis.
}

\details{
\tabular{ll}{
Package: \tab Compositional\cr
Type: \tab Package\cr
Version: \tab 4.7 \cr
Date: \tab 2021-05-13\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}, Giorgos Athineou <gioathineou@gmail.com> and Abdulaziz Alenazi <a.alenazi@nbu.edu.sa>.
}

\section{Maintainers }{
Michail Tsagris <mtsagris@uoc.gr>
}

\references{
Aitchison J. (1986). The statistical analysis of compositional data. Chapman & Hall.
}

\note{
Acknowledgments:

Michail Tsagris would like to express his acknowledgments to Professor Andy Wood and Dr Simon
Preston from the university of Nottingham for being his supervisors during his PhD in compositional data analysis.
We would also like to express our acknowledgments to Profesor Kurt Hornik (and also the rest of the R core team)
for his help with this package. Manos Papadakis, undergraduate student in the department of computer science,
university of Crete, is also acknowledged for his programming tips. Ermanno Affuso from the university of South Alabama
suggested that I have a default value in the "mkde"" function. Van Thang Hoang from Hasselt university spotted a bug in the
"js.compreg" function and is greatly acknowledged for that. Claudia Wehrhahn Cortes spotted a bug in the "diri.reg"" function
and she is greatly acknowledged for that. Philipp Kynast from Bruker Daltonik GmbH found a mistake in the function "mkde" which
is now fixed. Jasmine Heyse from the university of Ghent spotted a bug in the function "kl.compreg" which is now fixed.

}

\keyword{ Compositional data }
\keyword{ Regression }
\keyword{ Contour plots }
\keyword{ Discriminant analysis }
\keyword{ MLE }


